update codes and delete unnecessary dataset
Fix contents and structure, some expression.
Modify load data according to Nishat's share in the slack channel so Brian could run our notebook without downloading data.
Add Shirley's takeaway.
Had a discussion with Shirley about the evaluation matrics and we agreed that np.exp is not needed as features are not log transformed so don't need to add that, only keep the pipeline one.